dcs_lcd_color: Add shared RGB565 color helpers
Extract alpha_blend_rgb565, rgba8888_color_to_rgb565, and
related color helpers into a shared inline-header module.
Drop the unused struct Screen parameter.  Migrate st7789,
ili934x, and ili948x.
